pyproject.toml
[project]
name = "semantic-bridge-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Moti Elmakyes",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "python-socketio[client] (>=5.14.3,<6.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "chromadb (>=0.4.0,<1.0.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "mcp (>=1.1.0,<2.0.0)",
    "google-adk (>=1.18.0,<2.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "google-genai (>=1.50.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
